requirements.txt
# --- Core AI & LLM ---
langchain-google-genai>=1.0.1
langgraph>=0.0.30
pydantic>=2.6.4
python-dotenv>=1.0.1

# --- Databases ---
neo4j>=5.18.0
qdrant-client>=1.8.0

# --- Frontend & UI ---
streamlit>=1.32.0
pandas>=2.2.1
folium>=0.16.0
streamlit-folium>=0.18.0

# --- Utilities ---
requests>=2.31.0
numpy>=1.26.0